+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added experimental `non_blocking_lock_timeout` option
 - Added retries for `ANALYZE` when `lock_timeout_retries` is set
+- Fixed lock timeout retries after `safe_by_default` commits the transaction
 
 ## 2.8.0 (2026-05-14)